[reconfigurator] `BlueprintBuilder` cleanup 4/5 - when choosing a zpool, it must be from a disk present in the blueprint (#7106)
Builds and is staged on top of #7105.

The intended change here is in the first commit
(82741741926b589bf2d381e1dbbb8e435444871d): In
`BlueprintBuilder::sled_select_zpool()`, instead of only looking at the
`PlanningInput`, we also look at the disks present in the blueprint, and
only select a zpool that the planning input says is in service and that
we have in the blueprint.

This had a surprisingly-large blast radius in terms of tests - we had
_many_ tests which were adding zones (which implicitly selects a zpool)
from a `BlueprintBuilder` where there were no disks configured at all,
causing them to emit invalid blueprints. These should all be fixed as of
this PR, but I'm a little worried about test fragility in general,
particularly with an eye toward larger changes like #7078. Nothing to do
about that at the moment, but something to keep an eye on.

Fixes #7079.