Retry OxQL queries in integration tests
Adds a small `wait_for_condition()` loop to OxQL queries in integration
tests. This tries to catch the error where the query fails specifically
because the timeseries it names is not yet available, and return `None`
in these cases. We expect it to appear soon, so this isn't immediately
fatal.

Fixes #7154 and #7084.