[reconfigurator] BlueprintBuilder cleanup 5/5 - Convert discretionary ensure_zone_multiple_* to add_zone_* (#7107)
All of the callers of the `ensure_zone_multiple_*` methods to add
discretionary zones were, in practice, wanting to _add_ a specific
number of zones. This rewords all of those methods to be imperative,
allowing us to remove a pretty significant chunk of now-unnecessary
checks and logic.

Builds on and is staged on top of #7106.