Add disposition to blueprint disks (#7169)
This is the second PR related to fixing #7098.

BlueprintPhysicalDisksConfig is no longer an alias of
OmicronPhysicalDisksConfig, but instead wraps a new type
BlueprintPhysicalDiskConfig which itself wraps an
OmicronPhysicalDiskConfig and a BlueprintPhysicalDiskDisposition. This
change brings blueprint physical disks in line with blueprint datasets
and zones such that expungement is a first class notion in the
blueprint, and not implicit in the disk not being present in a
blueprint.

This change only adds the new types and makes them work with the
existing code. The underlying logic around expungement and
decommissioning has not changed. That will come in a follow up PR.