[reconfigurator] Fix DatasetsEditor internal zpool,kind -> ID map (#7216)
Blueprint zones don't contain explicit dataset IDs (yet; see #7214), so
`DatasetsEditor` maintained an internal cache mapping `(zpool, kind) ->
dataset ID`. However, that mapping is only unique for _in service_
datasets, and `DatasetsEditor` was erroneously trying to build it for
_all_ datasets (both in-service and expunged).

This PR adds a few property tests and fixes for the maintenance of this
cache, and should ensure that we only try to maintain the "at most one
dataset of a given kind on a given zpool" map for in-service datasets.