Cleanup oximeter collector internals (#7205)
- Move result sink tasks to their own module
- Move collection task innards to their own module, exposing a simpler
interface to the task via a new "handle" type. This makes the collection
message type private, and instead uses public methods on the handle for
dealing with messaging internally.
- Move the producer endpoint info from the task map into the task handle
itself, which simplifies a lot of the map manipulation in the agent
itself.
- Break up the massive `tokio::select!` in the collection task. This is
now in a main `run()` method on the new collection task type. That lets
us factor out each branch of the select (waiting for a timer to fire, or
a message from the oximeter agent, e.g.) into their own methods. This is
entirely NFC, it just makes everything much easier to understand and
debug.
- Closes #7202