omicron-package: work around cargo issue 8157 (#7218)
#5799 modified the `cargo build` command line omicron-package runs.
Previously it built up a list of packages to be built using the `-p`
flag; that PR changed it to use `--bin`. The goal was to build only the
binaries that are necessary for shipping; this avoids building
sled-agent-sim during releng, for instance.

We did not realize it at the time, but this invited the specter of
https://github.com/rust-lang/cargo/issues/8157 to wreak havoc; namely:

- Without `--package`, Cargo uses the `default-members` key of the
workspace Cargo.toml to determine which packages to build. `--bin` does
not cause the same thing to happen; saying `--bin` does _not_ imply
`--package [the package that the bin belongs to]`.
- `omicron-dev` belongs to `default-members` and has a normal dependency
on `nexus-test-utils`, which enables the `"testing"` feature of
`nexus-db-queries`.

https://github.com/oxidecomputer/omicron/issues/7208 is a known result
of this problem, but there might be more.

Fortunately the solution seems fairly easy, without reverting the
relevant changes from #5799: use _both_ `--package` and `--bin`. With
this change, the `"testing"` feature is no longer shown in the `cargo
build --unit-graph` and `nm target/release/nexus | demangle | grep
validate_volume_invar` no longer shows any matching testing-only
symbols.