Do not pass arbitrary strings to volume_create!
Accepting arbitrary strings is a bad idea! Instead, change
`volume_create` from accepting a `Volume` argument to accepting a
`VolumeUuid` and a `VolumeConstructionRequest` directly.

This wasn't a large lift in terms of changes, as most call sites of
`volume_create` pass in the result of `Volume::new`, which takes an id
and a serialized `VolumeConstructionRequest` string. This commit changes
that to skip performing the `serde_json::to_string` and pass the id and
volume construction request directly, and instead lets the
`volume_create` handle the serialization (this function already returns
a `Err(SerdeError)` variant, so this commit is not introducing anything
new).