Only perform region replacement for r/w regions
Region replacement does not work for read-only regions (there is a check
in the region_replacement_start saga that bails out if the supplied
region is read-only) - this is tracked by #6172. The current plan to
make read-only region replacement work is to use the same machinery as
region snapshot replacements (as they're both read-only), so this commit
changes the current query that finds expunged regions to replace to only
return read/write regions. This can be changed back in the future if the
plan for #6172 changes.