Actually check with the simulated Crucible agent
Query the simulated Crucible agent when performing a clean slate check
after the region snapshot replacement start saga tests.

This exposed that the original diff did not add the undo node to the
DAG, which this commit fixes.

Add a test for if the saga leaked the replacement region. This exposed
that the simulated Crucible agent would refuse to destroy Failed regions
"because the real agent would not let a Failed region be deleted", even
though the real one totally allows it.

Also adds a `test_action_failure_can_unwind` test for the same saga.