refine error messages about not having a propolis address
the chronology here seems to be: Nexus learned to talk directly to
Propolis to communicate with an instance's serial console, that
"instance to Propolis address" translation got outlined into
`propolis_addr_for_instance`, then for region and region snapshot
replacment sagas we needed to also talk to the relevant Propolis. at
that point, `propolis_client_for_instance` used
`propolis_addr_for_instance` too.

this is all fine, but `propolis_addr_for_instance` kept a few serial
console-specific error messages, which would show up in places like
region replacement if an instance is stopped or shut down at an
inopportune time.

in the process, i discovered that a request for an instance's serial
console history with incorrect parameters manifests as a 500 rather than
a 400. Propolis even returns a 400, but the failure to establish a
websocket console is presumed an internal error in Nexus, and so it's
wrapped as a 500 whose contents talk about a 400! fix that to just
return a 400.