live-tests: avoid relying on poorly-behaving DNS (#7267)
1. Live tests don't currently run because the check that you are on an
underlay network asks internal DNS for A/AAAA records for internal DNS
(via `hickory_resolver::Resolver::lookup_ip`). We don't have those, but
prior to #6308 we responded with whatever records we had for a given
name regardless of what type the query asked for. Live tests aren't run
in CI so we missed it.
2. `internal_dns_resolver::Resolver::lookup_ip` is now only used here
and in a unit test that is arguably doing something similar, so we
remove the function altogether.