[tests] make instance reincarnation tests less racy
A possible test flake exists in the tests for instance reincarnation.
This is caused by a race that occurs when the request to simulate an
instance's state transition is sent to the simulated sled-agent *before*
an instance-start saga sends the request to start the instance to that
sled-agent. When this occurs, the simulated state transition is lost,
and the test keeps waiting for it forever. See [this comment][1] for
details.

This commit resolves this by adding a new
`instance_wait_for_simulated_transition` helper, which is identical to
`instance_wait_for_state`, but with the addition of `instance_poke`
requests every time the instance is not observed to be in the desired
state. This is a bit of a blunt instrument, but it ensures that the
simulated sled-agent will always be told to simulate a state transition
after it's requested by the control plane.

Hopefully fixes #6727

[1]: https://github.com/oxidecomputer/omicron/issues/6727#issuecomment-2557807360