Put `BlueprintZoneConfig` into a map
This is a largely mechanical change to store zones in a Blueprint mapped
by their zone uuid rather than in a Vec. This is primarily to support
diffs of `BlueprintZonesConfig` via diffus. Diffus doesn't handle Vecs
in a manner where inserts, removals, and modifications are tracked based
on an id.

To minimize the PR diff, the zone id was not removed from
`BlueprintZoneConfig`. The key to the map must always match the value
in `BlueprintZoneConfig`. It's a shame a mismatch is now representable,
and we can go about trying to change that if necessary. We can also
just make it part of blippy for now. FWIW, this matches the pattern in
`BlueprintDatasetsConfig`.

Putting the `BlueprintZoneConfig` into a map means we no longer require
the sort method on `BlueprintZonesConfig` as the order is stable
now. However, we still want to ensure the same display sort order as
previously, so we make sure to sort things properly before creating
the table.

We'll want similar changes for `BlueprintPhysicalDisksConfig`. That can
come in a follow up PR.