[RSS] Initial blueprint: only set `address` for Crucible datasets (#7310)
This should prevent #7299 from occurring on new invocations of RSS; I'll
add a followup PR (with more testing) that adds the `CHECK` constraint
mentioned in that issue and a migration that does cleanup of any bad
dataset address rows.

While I was in here, fixed #7115. It was smaller than I thought;
should've done that a while ago.