evaluate subscription globs on creation
this way, rather than transforming the globs into patterns that we can
match in the database every time an event is dispatched, we instead
use globs to generate "exact" subscriptions to specific named event
classes. this avoids doing a full scan over the subscription table every
time an event is dispatched, since we can look up subscribed receivers
by event class name instead.

we'll have to do this processing on receiver creation or when a
subscription is added, and when new event classes are added (e.g. on
software updates).

thanks @andrewjstone for pointing me in this direction!