[reconfigurator] Introduce `IdMap` and use it in `BlueprintZonesConfig` (#7327)
@andrewjstone mentioned being bothered by our `BTreeMap<Id,
ValueWithAnId>` maps on #7315:

> The key to the map must always match the value in BlueprintZoneConfig.
It's a shame a mismatch is now representable, and we can go about trying
to change that if necessary. We can also just make it part of blippy for
now. FWIW, this matches the pattern in BlueprintDatasetsConfig.

This has bugged me too, and adding a blippy check crossed my mind. But I
think it would be better to make this just not possible, so this PR
introduces a newtype wrapper around `BTreeMap<T::Id, T>` that enforces
"key must be value's ID" via a mix of compile time API constraints
(e.g., `insert` only takes the value and generates the key on its own)
and runtime checks (e.g., when mutating a value stored in the map, it
will panic if that mutation changes the ID of the value).

The new code is all in `id_map.rs`, and the rest of the changes are
replacing the `BlueprintZonesConfig` map with it. If this looks
reasonable to folks, followup PRs to change the other similar
`BTreeMap`s to use `IdMap` should be straightforward.