Add a Diff Visitor for BlueprintZoneConfigs (#7336)
This visitor walks a diffus generated diff and provides callbacks for
users to hook into the parts of the diff that they care about.

To keep things minimal, we don't provide callback trait methods for
copies (when nothing changes), and we don't provide methods for most
diffus_derive generated types (the ones that start with `Edited`).

We anticipate composing these visitors to build up a full diff of a
Blueprint. They can also be used individually as demonstrated with the
included tests.