[support-bundles] Re-assign bundles to 'ourselves' from Nexus (#7351)
Slightly simplify the process of handling bundle reassignment.

Support bundles are "owned" by a single Nexus, for the purposes of
allocation of deletion. During blueprint execution, we look for expunged
Nexuses, and replace all their corresponding bundles with a valid Nexus.

Previously, we'd consider "one of the in-service Nexus zones" as the
owner.
In this PR, we take responsibility by using our *own* Nexus UUID as the
owner.

This has a slightly better chance of being valid (we're running), and it
simplifies the code by guaranteeing that there is a Nexus available to
make the replacement.