Set an arbitrary number of sled agents for tests (#7353)
As part of testing for oxidecomputer/crucible#1594, I needed the ability
to specify multiple extra sled agents. I pulled some work out of the fix
for that issue into this commit, which adds the `extra_sled_agents`
parameter to the `nexus_test` macro (also refactoring that code so that
other parameters can be added easier in the future).

Previously, `ControlPlaneTestContext` would configure two sled agents,
and later the impl of nexus_test_interface::NexusServer for Nexus would
mark the second sled agent as non-provisionable. This commit changes the
default number of sled agents created as part of
`ControlPlaneTestContext` to 1, and if more are created they are not
marked non-provisionable.

Many tests do not require two sled agents, and those that do can specify
how many they need with the new the `extra_sled_agents` parameter.

A lot of other changes in this diff are mechanical: for example changing
from reaching into the internals of `ControlPlaneTestContext` to using
functions.