Always allocate regions to different sleds
When performing region allocation _again_ for a particular volume and
increasing the redundancy, the region allocation query was not excluding
the sleds that have already been used by the existing allocation as
candidates for the new region (note that this only matters when the
allocation strategy mandates distinct sleds!). This resulted in region
replacement and region snapshot replacement having the potential to
allocate the replacement region onto the same sled as an existing
region.

This commit fixes the region allocation query, and then fixes the
fallout: many tests of region replacement or region snapshot replacement
were not creating many simulated sled agents, and therefore could not
properly satisfy the fixed region allocation query. The
`extra_sled_agents` parameter added in #7353 was used, along with
changing the `DiskTestBuilder` to simulated one zpool on many sleds.

Changing these replacement tests revealed the next problem: because all
the simulated sleds share the same IP (::1) and range of ports for the
fake Crucible agents, each region allocation would have the same target,
and the volume construction request would look like:

    "target": [
      "[::1]:1100",
      "[::1]:1100",
      "[::1]:1100"
    ]

This is very confusing for the replacement routines!

We can't change the simulated sled's IPs: they need to bind to localhost
to listen for requests. Therefore this commit also adds the idea of
"sled index", and this sled index is used to make each range of Crucible
agent ports unique.

This lead straight to the next problem: the simulated Pantry assumes
that there is really only one simulated sled-agent, and this has all the
storage! This is no longer true in a post-#7353 world. Snapshots would
not be able to complete because the simulated Pantry would only find one
of the three regions when searching the first simulated sled-agent's
storage.

The real Pantry would construct a Crucible Upstairs, and delegate
snapshot requests to that, so this commit also adds a simulated Crucible
Upstairs that's aware of each of the simulated sled agent's `Storage`,
and can therefore fulfill snapshot requests. Centralizing the place
where all simulated sled agents (and Pantry servers!) are created makes
this easy: the simulated Upstairs can be aware of all of the `Storage`
objects, and the simulated Pantry can use that simulated Upstairs to
take fake snapshots.

Fixes oxidecomputer/crucible#1594.