Fix `test_sled_instance_list` (#7367)
`test_sled_instance_list` was broken by the most recent change to the
control plane test context: previously only one of the two sled agents
(the first one) was provisionable for instances, and the test would
check that one to test the sled instances list endpoint. The most recent
change to the test context made all sled agents provisionable but didn't
update the test to check all the sleds for the provisioned instance.
This resulted in a test flake: if the instance was allocated to the
first sled, it would pass. If it was allocated to the second sled, it
would time out.

Fixes #7365