[oxlog] skip unnecessary lstat (#7372)
We should not incur the cost of calling `lstat` on every log file oxlog
walks over when we are returning all log files regardless of size. If
someone requests that we don't show 0 length files through oxlog the
binary or the library then we will do the `lstat` in order to check the
file length.

I originally found this edge case when operating on a sled with 600k
plus log files that were being walked. Figuring out what we want to do
about aging out old debug logs is out of the scope of this PR.