WIP
Make current inventory stuff MGS-specific

Transceivers are plumbed from dpd to wicket

BUT: this approach cannot work. By talking to dpd, we only get the
transceivers in the switch we happen to be plugged into. We also need to
get those from the other switch. MGS can do that because it's on the
mgmt network. dpd cannot directly, but we can by using the transceiver
controller crate ourselves. Need to think about whether that's good, but
we could do it in theory.