A VMM in state 'starting' can receive requests
If a VMM is in state 'starting' and cannot activate a given disk, it
will hang. Region replacement or region snapshot replacement may alter
the VCR of that disk to one that *could* be activated, but the current
drive saga will not attempt to send requests to a VMM in 'starting'. Fix
this - any time a Propolis is expected to be there, it should be ok to
receive these requests. Otherwise a repair can be stuck, and the VMM
could fail to stop if it gets stuck deactivating too!