Change `dataset` table to only hold Crucible datasets (#7386)
The diff of this PR is large but _very_ mechanical. The actual changes
here are:

* The `dataset` table has been renamed `crucible_dataset` (and
correspondingly, the `Dataset` model struct is now `CrucibleDataset`,
and `Datastore::dataset_*` methods are now
`Datastore::crucible_dataset_*`)
* The `zone_name` column has been dropped (it was always NULL for
Crucible datasets, enforced by a `CHECK` constraint)
* The `quota`, `reservation`, and `compression` columns have been
dropped (they were always NULL for Crucible datasets, although this was
not enforced by a `CHECK` constraint; this information is kept in the
blueprint and is not read from this table)
* The `ip`, `port`, and `size_used` columns now have `NOT NULL`
constraints (they already had `NOT NULL` constraints specifically for
Crucible datasets via a `CHECK` constraint; correspondingly, the
`CrucibleDataset` model type has non-optional fields for these columns)

and everything else is followup to account for this. Most of that
followup is just renaming things. A few nontrivial exceptions:

* There were a handful of places that had to deal with the optionality
of `address` by failing if it wasn't set for Crucible datasets; that
error handling is now gone since `address` is no longer optional.
* Some tests specifically operated on the `dataset` table using a mix of
Crucible and non-Crucible datasets. I tried to update these as
reasonably as I could while keeping the spirit of the test in tact, but
a closer look at these bits would be good.
* The rack init request in the RSS -> Nexus handoff now only includes
Crucible datasets.

I think this gets us most of the way toward `crucible_dataset` being a
"rendezvous table" in the RFD 540 sense. As of this PR, the blueprint
executor task is still responsible for adding entries to this table, but
I'll move that to the rendezvous background task in a followup.

Fixes #7299. Gets us halfway to #7301 by fixing the `dataset` side of
that issue (`bp_omicron_dataset` still needs work).